requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
asyncua>=1.1.0
aiomqtt>=2.0.0
pyyaml>=6.0
pydantic>=2.0
websockets>=12.0
httpx>=0.27.0
